Refactor the GS notices to share actions between different notices
