+++ projects/packages/jetpack-mu-wpcom/src/features/wpcom-global-styles/notice.scss
@@ -15,6 +15,15 @@
 	.components-notice__content {
 		margin-right: 0;
 	}
+
+	.components-notice__action.components-button {
+		margin-left: 0;
+		margin-top: 12px;
+
+		// Protect the button against long strings overflowing the sidebar
+		text-align: left;
+		white-space: normal;
+	}
 }
 
 // Inline the action buttons when there is another pinned notice,
@@ -46,7 +55,7 @@
 	}
 }
 
-.components-notice-list .components-notice__action.components-button.wpcom-global-styles-action-has-icon {
+.components-notice__action.components-button.wpcom-global-styles-action-has-icon {
 	display: flex;
 
 	&::after,
